Add files via upload
DWT works on the 2023 L1 signal, I have attached the text file along for testing. It wasn't very noisy so you will have to zoom in to see the fluctuations about the denoised wave. It also uses STFT to produce 20 Hz bands, which may be a step in the right direction.